follow best practices and improve the security of the application
